feat: add financial page and remove stats from other pages
